** Corrected issues **
- Fix several minor bugs (from the auto generated report)
- Fixed lexer to correctly read end of statement again (was broken at the last commit)
- Corrected issues : #32, #29, #33, #27, #36, #35, #32,
- Now correctly handles errors if 3P failed to write in userDefineLang.xml (which would cause 3P to fail loading)
- #41 : adding support for the key word EDITING (code folding + correct block indentation
- #39 : you can now open the main application window from any file (instead of only progress files before)

+++ 3PA/MainFeatures/ErrorHandler.cs
@@ -51,7 +51,7 @@ namespace _3PA.MainFeatures {
         /// <param name="fileName"></param>
         public static void ShowErrors(Exception e, string message, string fileName) {
             Log(e.ToString());
-            MessageBox.Show(@"Attention user! An error has occured while loading in the following file :" + "\n\n"
+            MessageBox.Show(@"Attention user! An error has occurred while loading in the following file :" + "\n\n"
                 + fileName +
                 "\n\n" + @"The file has been suffixed with '_errors' to avoid further problems.", AssemblyInfo.ProductTitle + " error message", MessageBoxButtons.OK, MessageBoxIcon.Error);
             if (File.Exists(fileName + "_errors"))
@@ -77,7 +77,7 @@ namespace _3PA.MainFeatures {
                 Task.Factory.StartNew(() => {
                     try {
                         if (Config.Instance.LogError) {
-                            if (UserCommunication.SendIssue(File.ReadAllText(PathErrorToSend), Config.SendLogUrl)) {
+                            if (!Config.Instance.GlobalDontAutoPostLog && UserCommunication.SendIssue(File.ReadAllText(PathErrorToSend), Config.SendLogUrl)) {
                                 if (File.Exists(PathErrorToSend))
                                     File.Delete(PathErrorToSend);
                             }
@@ -94,7 +94,7 @@ namespace _3PA.MainFeatures {
                     UserCommunication.Notify("The last action you started has triggered an error and has been cancelled.<br><br>1. If you didn't ask anything from 3P then you can probably ignore this message and go on with your work.<br>2. Otherwise, you might want to check out the error log below :" +
                         (File.Exists(PathErrorfile) ? "<br><a href='" + PathErrorfile + "'>Link to the error log</a>" : "") +
                         "<br>Consider opening an issue on GitHub :<br><a href='https://github.com/jcaillon/3P/issues'>https://github.com/jcaillon/3P/issues</a>" + "<br><br><b>Level 0 support : restart Notepad++ and see if things are getting better!</b>",
-                        MessageImg.MsgPoison, "An error has occured", message,
+                        MessageImg.MsgPoison, "An error has occurred", message,
                         args => {
                             Npp.Goto(args.Link);
                             args.Handled = true;
@@ -102,7 +102,7 @@ namespace _3PA.MainFeatures {
                         0, 500);
                 else
                     UserCommunication.Notify("The last action you started has triggered an error and has been cancelled.<br>If you didn't ask anything from 3P then you can probably ignore this message and go on with your work.<br>Otherwise, another try will probably fail as well.<br>Consider restarting Notepad++ as it might solve this problem.<br>Finally, you can use the link below to open an issue on GitHub and thus help programmers debugging 3P :<br><a href='https://github.com/jcaillon/3P/issues'>https://github.com/jcaillon/3P/issues</a>",
-                        MessageImg.MsgPoison, "An error has occured", message,
+                        MessageImg.MsgPoison, "An error has occurred", message,
                         args => {
                             Npp.Goto(args.Link);
                             args.Handled = true;
@@ -111,7 +111,7 @@ namespace _3PA.MainFeatures {
             } catch (Exception x) {
                 DirtyLog(x);
                 // display the error message the old way
-                MessageBox.Show("An unidentified error has occured, probably while loading the plugin.\n\nThere is a hugh probability that it will cause the plugin to not operate normally.\n\nTry to restart Notepad++, consider opening an issue on : https://github.com/jcaillon/3P/issues if the problem persists.", AssemblyInfo.ProductTitle + " error message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                MessageBox.Show("An unidentified error has occurred, probably while loading the plugin.\n\nThere is a hugh probability that it will cause the plugin to not operate normally.\n\nTry to restart Notepad++, consider opening an issue on : https://github.com/jcaillon/3P/issues if the problem persists.", AssemblyInfo.ProductTitle + " error message", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
         }
